drivers: virtio: add IVSHMEM to virtio-mmio driver
This patch adds the IVSHMEM as a memory area option
to be utilized by the virtio-mmio device driver, it
overcomes some limitations of the Hypervisorless
Virtio-mmio from openamp by doing:

- Add an extra option to tell the virtio-mmio driver
we are about to use IVSHMEM, this overrides the de
fault driver behavior and allows to use IVSHMEM offsets
instead of absolute address

- Modify the virtio_mmio_init driver function to make
possible to resolve the virtio-mmio cfg and shmem addresses
to be resolved in runtime.

- Adds an extra IVSHMEM heap area to allocate driver side
Vrings, this overcomes other limitation from Hypervisorless
virtio-mmio where a linker special section was needed to place
the created VRINGS, since IVSHMEM addresses are resolved in
runtime, we needd an special heap in the same space of the
bounce buffer area allowing to allocate the shared VRINGS
from there, the zephyr heap subsystem was used as base
to create an runtime VRING allocator for IVSHMEM area.

- Change the driver priority order, since the IVSHMEM
driver from Zephyr is setup in POST_KERNEL priority we needed
to decrease the priority of virtio-mmio driver in order to ensure
its initialization AFTER the IVSHMEM driver and IPM.

Signed-off-by: Felipe Neves <felipe.neves@linaro.org>